LayoutTests/fast/masking/parsing-clip-path-iri-expected.txt

Test clip-path IRIs

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS innerStyle("-webkit-clip-path", "url(#clip1)") is "url(#clip1)"
PASS innerStyle("-webkit-clip-path", "url(clip.svg#clip1)") is "url(clip.svg#clip1)"
PASS computedStyle("-webkit-clip-path", "url(#clip1)") is "url(#clip1)"
PASS computedStyle("-webkit-clip-path", "url(clip.svg#clip1)") is "url(clip.svg#clip1)"
PASS successfullyParsed is true

TEST COMPLETE
